pyproject.toml
[project]
name = "short-url"
version = "0.1.0"
description = "Сервис сокращения ссылок (аналог bitly)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "loguru>=0.7.3",
    "pybase62>=1.0.0",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
